refactor: :recycle: выделить договоры парных модулей
Нейтральные типовые договоры разрывают взаимные зависимости реализаций. Прежние точки экспорта сохранены для совместимости потребителей.
