+++ packages/core/metadata/appliedObjects/configuration/syncStateContracts.ts
@@ -1,0 +1,10 @@
+export interface XmlSyncState {
+  version: 1
+  files: Record<string, string>
+}
+
+export interface XmlSyncStateDiff {
+  added: string[]
+  changed: string[]
+  deleted: string[]
+}
