+++ packages/core/metadata/validation/dataPath/formatContracts.ts
@@ -1,0 +1,1 @@
+export type DataPathFormatDirection = "internal-to-yaml" | "yaml-to-internal"
