+++ packages/core/xml/import/contracts.ts
@@ -1,0 +1,4 @@
+export type ImportContentFromXMLOptions = {
+  preserveXsiNil?: true
+  preserveEmptyElements?: true
+}
